requirements.txt
python-telegram-bot==20.7
apscheduler==3.11.0
requests==2.32.4
pytz==2025.2
